requirements.txt
pytest
pdfminer.six
opencv-python
pyzbar
pdf2image
PyMuPDF
